python-analytics/requirements.txt
aiokafka==0.12.0
PyYAML